imsz
Enum ImFormat

#[non_exhaustive]
pub enum ImFormat {

    GIF = 1,
    PNG = 2,
    BMP = 3,
    JPEG = 4,
    WEBP = 5,
    QOI = 6,
    PSD = 7,
    XCF = 8,
    ICO = 9,
    AVIF = 10,
    TIFF = 11,
    OpenEXR = 12,
    PCX = 13,
    TGA = 14,
    DDS = 15,
    HEIF = 16,
    JP2K = 17,
    DIB = 18,
    VTF = 19,
    ILBM = 20,

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

GIF = 1

Graphics Interchange Format files in version GIF87a or GIF89a.

PNG = 2

Portable Network Graphics files. Requires the first chunk to be IHDR.

BMP = 3

Windows Bitmap, both for Windows 2.0 (BITMAPCOREHEADER) and for newer versions (BITMAPINFOHEADER).

JPEG = 4

Joint Photographic Experts Group files.

WEBP = 5

WebP files. Supported sub-formats: VP8 , VP8L, VP8X.

QOI = 6

Quite OK Image format files.

PSD = 7

Adobe Photoshop files.

XCF = 8

GIMP files.

ICO = 9

ICO files can contain multiple images. This returns the dimensions of the biggest image in the file.

AVIF = 10

AV1 Image File Format.

TIFF = 11

Tag Image File Format. Supports big endian and little endian TIFF files.

OpenEXR = 12

OpenEXR files.

PCX = 13

PiCture eXchange files.

TGA = 14

TARGA (Truevision Advanced Raster Graphics Adapter) files.

Only if the file ends in b"TRUEVISION-XFILE.\0" since otherwise there is no good way to detect TGA files. Note that this string is optional to this file format and thus there can be TGA files that arenโt supported by this library.

DDS = 15

DirectDraw Surface files.

HEIF = 16

HEIC/HEIF files. These are extremely similar to AVIF and use the same parsing code.

JP2K = 17

JPEG 2000 files.

DIB = 18

Device-Independent Bitmap files.

VTF = 19

Valve Texture Format.

ILBM = 20

Interleaved Bitmap files, including Planar Bitmap variant.

Implementations

impl ImFormat

pub const fn name(&self) -> &'static str